pyproject.toml
[project]
name = "fastapi-team-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich==0.9.1",
    "asyncpg==0.30.0",
    "coverage==7.10.5",
    "fastapi==0.116.1",
    "google-generativeai>=0.8.5",
    "mypy==1.17.1",
    "python-dotenv==1.1.1",
    "tortoise-orm==0.25.1",
    "uvicorn==0.35.0",
]

[dependency-groups]
dev = [
    "autoflake==2.3.1",
    "black==25.1.0",
    "flake8==7.3.0",
    "isort==6.0.1",
    "pre-commit==4.3.0",
    "pytest==8.4.1",
    "ruff==0.12.10",
]

[tool.ruff]
line-length = 120

[tool.flake8]
max-line-length = 120

[tool.aerich]
tortoise_orm = "core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
